refactor modal out of link control component
